Build extension from modules using webpack
Our code uses modules to allow import/export (in particular, for tests).
Chrome extensions don't actually support that;
for example, see https://stackoverflow.com/q/48104433
Instead, all the background scripts are imported into the same namespace.

This commit goes all-in on modules:
it adds webpack to (intelligently) concatenate all the modules in the
source code into a single background script, which becomes the only script
that gets loaded in the extension.

As a side effect, this commit does away with the `extension` directory.
Instead, the extension files are split across the `src` code directory
and `assets`. Running `npm build` generates the `chrome_extension`
directory, which holds the completed extension.