Fix bug in RegExp usage
We reuse a global RegExp object to execute our searches.
However, RegExp is stateful, so it would retain state between the different
invocations of the exec() function, even if the string being searched
had changed in the meantime.
This results in failed searches, because the RegExp would skip over
matches at the beginning of the string, so that it could begin at
lastIndex.

For a more worked-out example of how this happens,
see https://stackoverflow.com/q/1520800

This bug was especially subtle for us, because the searches in this case
failed silently (returned empty array, no matches) rather than throwing
errors.

The fix in this commit is to manually reset the search state at the
beginning of searching a new string.