Remove lazy evaluation in GoogleInteractions
This fixes a regression introduced in 1aa529a, when interaction fetching
and processing was moved into the background script. Afterwards,
interactions are sent to the content-script over postMessage, which
implicitly serializes the Interaction objects as JSON. However, since
the GoogleInteraction objects relied on accessors, the object that was
received ended up missing all the important fields.

This commit implements a fix (albeit a slightly hacky one).
It gets rid of the accessors and just sets the fields on the
GoogleInteraction object, so that it will (eventually) be properly
serialized.