Change when interactions are randomized
Interactions shown to participants are randomly sampled.
Up to now, the random sampling was done at selection-time.
In this commit, the interactions are randomly shuffled during the
validation phase, and at selection-time we simply choose the next unused
interaction.

(We still do a check in real-time to make sure it's a valid interaction.)

The motivation for this is upcoming changes to prioritize interactions
with recordings.

This commit also fixes a bug in the logic about "seen" interactions.
We keep track of these because sometimes the webpage re-requests an
"old" question. (It's always the currently displayed one; this happens
when form validation fails and Qualtrics re-renders the same page.)
To allow for this, we keep track of the interactions that have been
displayed. However, "bad" interactions (that we don't want to use) got
added to it as well, which meant the indexing was off. i.e., when the
extension went to get the last seen interactions, it may actually have
been returning one of the recently bypassed bad ones.